backend/requirements.txt
alembic==1.9.2
Flask==2.1.0
Flask-Login==0.6.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.8.6
requests==2.25.0
oauthlib==3.1.0
SQLAlchemy==1.3.23
pytest==6.2.1
pytest-mock==3.10.0
web3==5.28.0
py-solc-x==1.1.1
eth-tester==0.8.0b3
dwollav2==2.2.1